requirements.txt
streamlit==1.44.1
streamlit-autorefresh==1.0.1
pyserial==3.5
scapy==2.5.0
cryptography==44.0.2
pandas==2.2.3
PyYAML==6.0.2
plotly==5.24.1
